Removed unused old files after refactor - replaced references in main.cpp of RenderDevice to WorkGraphsDXR
