--- src/ascend/RayTracingRHI.cpp
@@ -1,80 +1,0 @@
-#include "RayTracingRHI.h"
-
-RayTracingRHI::RayTracingRHI()
-{
-
-}
-RayTracingRHI::~RayTracingRHI()
-{
-
-}
-
-void RayTracingRHI::Initialize()
-{
-
-}
-
-void RayTracingRHI::CreateRaytracingInterfaces()
-{
-
-}
-
-void RayTracingRHI::CreateRootSignatures()
-{
-
-}
-
-void RayTracingRHI::CreateRaytracingPSO()
-{
-
-}
-
-void RayTracingRHI::CreateDescriptorHeap()
-{
-
-}
-
-void RayTracingRHI::BuildAccelerationStructures()
-{
-
-}
-
-void RayTracingRHI::BuildShaderTables()
-{
-
-}
-
-void RayTracingRHI::CreateRaytracingOutputResource()
-{
-
-}
-
-void RayTracingRHI::CreateLocalRootSignatureSubobjects(CD3DX12_STATE_OBJECT_DESC* raytracingPipeline)
-{
-
-}
-
-void RayTracingRHI::CreateDeviceDependentResources()
-{
-
-}
-
-void RayTracingRHI::CreateWindowSizeDependentResources()
-{
-
-}
-
-void RayTracingRHI::DoRaytracing()
-{
-
-}
-
-void RayTracingRHI::CopyRaytracingOutputToBackbuffer()
-{
-
-}
-
-UINT RayTracingRHI::AllocateDescriptor(D3D12_CPU_DESCRIPTOR_HANDLE* cpuDescriptor, UINT descriptorIndexToUse)
-{
-	return 0;
-}
