--- src/ascend/WindowsApplication.cpp
@@ -1,73 +1,0 @@
-#include "WindowsApplication.h"
-#include "WinUser.h"
-
-HWND WindowsApplication::m_hwnd = nullptr;
-//Renderer* render = nullptr;
-/*
-int WindowsApplication::Run(HINSTANCE hInstance, int nCmdShow)
-{
-	WNDCLASSEX windowClass = { 0 };
-	windowClass.cbSize = sizeof(WNDCLASSEX);
-	windowClass.style = CS_HREDRAW | CS_VREDRAW;
-	windowClass.lpfnWndProc = WindowProc;
-	windowClass.hInstance = hInstance;
-	windowClass.hCursor = LoadCursor(NULL, IDC_ARROW);
-	windowClass.lpszClassName = L"ascendClass";
-	RegisterClassEx(&windowClass);
-
-	RECT windowRect = { 0, 0, 800, 800 };
-	AdjustWindowRect(&windowRect, WS_OVERLAPPEDWINDOW, FALSE);
-
-    // Create the window and store a handle to it.
-    m_hwnd = CreateWindow(
-        windowClass.lpszClassName,
-        L"ascend.editor",
-        WS_OVERLAPPEDWINDOW,
-        CW_USEDEFAULT,
-        CW_USEDEFAULT,
-        windowRect.right - windowRect.left,
-        windowRect.bottom - windowRect.top,
-        nullptr,        // We have no parent window.
-        nullptr,        // We aren't using menus.
-        hInstance,
-        nullptr);
-    //render = new Renderer();
-    //render->Initialize();
-    //ShowWindow(m_hwnd, nCmdShow);
-
-    // Main
-        // Main sample loop.
-    MSG msg = {};
-    while (msg.message != WM_QUIT)
-    {
-        // Process any messages in the queue.
-        if (PeekMessage(&msg, NULL, 0, 0, PM_REMOVE))
-        {
-            TranslateMessage(&msg);
-            DispatchMessage(&msg);
-        }
-    }
-
-    // Return this part of the WM_QUIT message to Windows.
-    return static_cast<char>(msg.wParam);;
-}
-
-LRESULT CALLBACK WindowsApplication::WindowProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
-{
-    switch (message)
-    {
-    case WM_DESTROY:
-        render->Shutdown();
-        PostQuitMessage(0);
-        return 0;
-    case WM_PAINT:
-        render->Update();
-    }
-
-    extern IMGUI_IMPL_API LRESULT ImGui_ImplWin32_WndProcHandler(HWND hWnd, UINT msg, WPARAM wParam, LPARAM lParam);
-    if (ImGui_ImplWin32_WndProcHandler(hWnd, message, wParam, lParam))
-        return true;
-    // Handle any messages the switch statement didn't.
-    return DefWindowProc(hWnd, message, wParam, lParam);
-}
-*/
